Move Slack actions to separate package and split in multiple files
