+++ blossom/api/views/submission.py
@@ -40,11 +40,7 @@ from blossom.api.models import Source, Submission, Transcription, TranscriptionC
 from blossom.api.pagination import StandardResultsSetPagination
 from blossom.api.serializers import SubmissionSerializer
 from blossom.api.slack import client as slack
-from blossom.api.slack.actions import (
-    ReportMessageStatus,
-    ask_about_removing_post,
-    update_submission_report,
-)
+from blossom.api.slack.actions.report import ask_about_removing_post, update_submission_report, ReportMessageStatus
 from blossom.api.slack.transcription_check.messages import send_check_message
 from blossom.api.views.volunteer import VolunteerViewSet
 from blossom.authentication.models import BlossomUser
